+++ src/js/polyfill/ForEach.polyfill.js
@@ -1,0 +1,12 @@
+/**
+ * ForEach polyfill
+ */
+export default function setUpForEach(){
+    if(window.NodeList && !NodeList.prototype.forEach) {
+        NodeList.prototype.forEach = Array.prototype.forEach;
+    }
+    
+    if(window.HTMLCollection && !HTMLCollection.prototype.forEach) {
+        HTMLCollection.prototype.forEach = Array.prototype.forEach;
+    }
+}