+++ src/js/polyfill/ObjectEntries.polyfill.js
@@ -1,0 +1,10 @@
+/**
+ * Object.entries polyfill
+ */
+export default function setObjectEntries(){
+    if(!Object.entries){
+        Object.entries = (obj) => Object.keys(obj).map(key => {
+            return [key, obj[key]];
+        });
+    }
+}